+++ package.json
@@ -59,7 +59,6 @@
     "semver": "7.3.5",
     "set-cookie-parser": "2.4.8",
     "source-map-support": "0.5.19",
-    "typescript": "4.3.5",
     "ws": "8.2.0"
   },
   "optionalDependencies": {
@@ -93,7 +92,8 @@
     "ts-jest": "27.0.5",
     "ts-node": "10.2.1",
     "typedoc": "0.21.6",
-    "uuid": "8.3.2"
+    "uuid": "8.3.2",
+    "typescript": "4.3.5"
   },
   "funding": [
     {
